gptscripter/tools/applescript_runner/requirements.txt
griptape-core>=0.8.0